+++ core/src/main/java/com/sap/cloud/security/token/validation/validators/XsuaaJwtIssuerValidator.java
@@ -1,7 +1,7 @@
 package com.sap.cloud.security.token.validation.validators;
 
-import static com.sap.cloud.security.core.Assertions.assertNotEmpty;
 import static com.sap.cloud.security.token.validation.ValidationResults.*;
+import static com.sap.cloud.security.xsuaa.Assertions.assertHasText;
 
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -26,7 +26,7 @@ public class XsuaaJwtIssuerValidator implements Validator<Token> {
 	 *            {@link OAuth2ServiceConfiguration#getDomain()}
 	 */
 	public XsuaaJwtIssuerValidator(String uaaDomain) {
-		assertNotEmpty(uaaDomain, "uaaDomain must not be null or empty.");
+		assertHasText(uaaDomain, "uaaDomain must not be null or empty.");
 		this.domain = uaaDomain;
 	}