+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/client/RequestParameterBuilder.java
@@ -48,6 +48,11 @@ class RequestParameterBuilder {
 		return this;
 	}
 
+	public RequestParameterBuilder withToken(String token) {
+		parameters.put(ASSERTION, token);
+		return this;
+	}
+
 	public Map<String, String> buildAsMap() {
 		return parameters;
 	}