+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.3.0
+- Issue: Spring tests fail with version `2.2.0`, when auto-configuration is disabled and no `RestOperations` bean is specified.
+- [token-client] Supports basically JWT Bearer Token Grant as documented [here](https://docs.cloudfoundry.org/api/uaa/version/74.4.0/index.html#jwt-bearer-token-grant).
+- [token-client] Supports the setup of mTLS with X509 client certificates. Use [`SSLContextFactory`](/token-client/src/main/java/com/sap/cloud/security/xsuaa/mtls/SSLContextFactory.java) to configure a `SSLContext` for Apache Rest Client to identify your client with the XSUAA identity service as documented [here](/token-client/README.md).
+
 ## 2.2.0
 - [spring-xsuaa] `PropertySourceFactory` supports custom property sources and default can optionally be disabled with  `spring.xsuaa.disable-default-property-source=true`
 - [spring-xsuaa] Supports Spring Core `5.2.0.RELEASE`and Spring Boot `2.2.0.RELEASE`